refactor(http): Re-home core-backed runtime glue
Move the remaining NodeClientCore-backed HTTP bootstrap adapters into runtime-owned endpoint packages.

Keep SimpleToadletServer's setCore/getCore compatibility behavior intact, make the narrow HTTP seams public only for runtime-owned adapters, and move the implementation-specific tests to the new runtime HTTP packages.
