+++ src/main/java/network/crypta/runtime/endpoints/http/CoreFProxyRuntimeSupport.java
@@ -1,8 +1,10 @@
-package network.crypta.clients.http;
+package network.crypta.runtime.endpoints.http;
 
 import java.io.File;
 import java.util.Objects;
 import network.crypta.client.async.ClientContext;
+import network.crypta.clients.http.FProxyRuntimeSupport;
+import network.crypta.clients.http.FProxyToadlet;
 import network.crypta.config.SubConfig;
 import network.crypta.node.NodeClientCore;
 import network.crypta.node.SecurityLevels.NETWORK_THREAT_LEVEL;
@@ -11,11 +13,12 @@ import network.crypta.node.SecurityLevels.PHYSICAL_THREAT_LEVEL;
 /**
  * Core-backed implementation of {@link FProxyRuntimeSupport}.
  *
- * <p>This adapter is the HTTP package's assembly seam between {@link FProxyToadlet} and {@link
+ * <p>This adapter is the runtime-owned HTTP bootstrap seam between {@link FProxyToadlet} and {@link
  * NodeClientCore}. It translates the broad daemon API into the small set of values and callbacks
  * that FProxy currently consumes, including threat-level state, download policy, and the node
  * executor used for background follow-up work. The adapter stays package-private because it is an
- * internal wiring detail of {@code clients/http}, not a reusable cross-module contract.
+ * internal wiring detail of runtime-owned HTTP endpoint bootstrap, not a reusable cross-module
+ * contract.
  *
  * <p>The record is immutable after construction and delegates every call directly to the wrapped
  * core. It does not cache mutable security state or configuration snapshots, so callers continue to
